+++ app/src/free/kotlin/ac/mdiq/podcini/playback/cast/CastEnabledActivity.kt
@@ -5,8 +5,7 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.compose.runtime.Composable
 
 /**
- * Activity that allows for showing the MediaRouter button whenever there's a cast device in the
- * network.
+ * Activity that allows for showing the MediaRouter button whenever there's a cast device in the network.
  */
 abstract class CastEnabledActivity : AppCompatActivity() {
     val TAG = this::class.simpleName ?: "Anonymous"