+++ app/src/play/kotlin/ac/mdiq/podcini/playback/cast/CastEnabledActivity.kt
@@ -22,8 +22,7 @@ import com.google.android.gms.common.ConnectionResult
 import com.google.android.gms.common.GoogleApiAvailability
 
 /**
- * Activity that allows for showing the MediaRouter button whenever there's a cast device in the
- * network.
+ * Activity that allows for showing the MediaRouter button whenever there's a cast device in the network.
  */
 abstract class CastEnabledActivity : AppCompatActivity() {
     private var canCast by mutableStateOf(false)